#978 move group sematic types from ch02 to ch04
