+++ docs/chapter_blurbs.md
@@ -5,7 +5,7 @@ What does each one do?
 
 - **ch00_purpose**: Introduces why Jaar has been created.
 - **ch01_data_toolbox**: Creates boring tools for computer data manipulation.
-- **ch02_rope_logic**: Defines Term Classes: Knots, Labels, RopeTerms, FaceNames, GroupNames, MomentLabels
+- **ch02_rope_logic**: Defines Term Classes: Knots, Labels, RopeTerms, GroupNames, MomentLabels
 - **ch03_finance_logic**: Defines tools for financial allotment to ledgers.
 - **ch04_group_logic**: Defines debt and cred to Voices and Memberships to Groups.
 - **ch05_reason_logic**: Defines ReasonUnits, FactUnits, and Facts decide if a Reason status is True
