+++ src/ch02_rope_logic/_ref/ch02_ref.json
@@ -1,5 +1,5 @@
 {
-  "chapter_blurb": "Defines Term Classes: Knots, Labels, RopeTerms, FaceNames, GroupNames, MomentLabels",
+  "chapter_blurb": "Defines Term Classes: Knots, Labels, RopeTerms, GroupNames, MomentLabels",
   "chapter_content": "Defines what a Rope is and required format for groups names and individual names.",
   "chapter_description": "ch02_rope_logic",
   "chapter_number": 2
