+++ src/ch04_group_logic/group.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from src.ch01_data_toolbox.dict_toolbox import get_1_if_None, get_dict_from_json
-from src.ch02_rope_logic._ref.ch02_semantic_types import (
+from src.ch04_group_logic._ref.ch04_semantic_types import (
     GroupTitle,
     VoiceName,
     default_knot_if_None,
