+++ src/ch04_group_logic/test/test_z_labor_.py
@@ -1,9 +1,9 @@
-from src.ch02_rope_logic._ref.ch02_semantic_types import GroupTitle
 from src.ch04_group_logic._ref.ch04_keywords import (
     parent_solo_str,
     party_title_str,
     solo_str,
 )
+from src.ch04_group_logic._ref.ch04_semantic_types import GroupTitle
 from src.ch04_group_logic.group import groupunit_shop, membership_shop
 from src.ch04_group_logic.labor import (
     LaborHeir,
