+++ src/ch04_group_logic/test/test_z_labor_json.py
@@ -1,4 +1,4 @@
-from src.ch02_rope_logic._ref.ch02_semantic_types import GroupTitle
+from src.ch04_group_logic._ref.ch04_semantic_types import GroupTitle
 from src.ch04_group_logic._ref.ch04_keywords import party_title_str, solo_str
 from src.ch04_group_logic.labor import (
     laborunit_get_from_dict,
