+++ src/ch04_group_logic/voice.py
@@ -4,7 +4,7 @@ from src.ch01_data_toolbox.dict_toolbox import (
     get_1_if_None,
     get_dict_from_json,
 )
-from src.ch02_rope_logic._ref.ch02_semantic_types import VoiceName
+from src.ch04_group_logic._ref.ch04_semantic_types import VoiceName
 from src.ch02_rope_logic.rope import (
     default_knot_if_None,
     is_labelterm,
