+++ src/ch07_belief_logic/belief_tool.py
@@ -1,5 +1,5 @@
 from src.ch01_data_toolbox.dict_toolbox import create_csv
-from src.ch02_rope_logic._ref.ch02_semantic_types import (
+from src.ch07_belief_logic._ref.ch07_semantic_types import (
     MomentLabel,
     RopeTerm,
     VoiceName,
