+++ src/ch10_pack_logic/test/_util/ch10_examples.py
@@ -1,4 +1,4 @@
-from src.ch02_rope_logic._ref.ch02_semantic_types import MomentLabel
+from src.ch07_belief_logic._ref.ch07_semantic_types import MomentLabel
 from src.ch02_rope_logic.rope import create_rope
 from src.ch09_belief_atom_logic.atom_main import BeliefAtom, beliefatom_shop
 from src.ch10_pack_logic._ref.ch10_keywords import (
